requirements.txt
fastapi==0.118.0
langgraph==0.6.8
openai==1.109.1
pydantic==2.11.9
python-dotenv==1.1.1
Requests==2.32.5
uvicorn==0.37.0
